+++ packages/orq-rc/src/orq_ai_sdk/models/createpromptop.py
@@ -101,7 +101,7 @@ r"""The version of photoReal to use. Must be v1 or v2. Only available for `leona
 EncodingFormat = Literal["float", "base64"]
 r"""The format to return the embeddings"""
 
-ReasoningEffort = Literal["low", "medium", "high"]
+ReasoningEffort = Literal["disable", "low", "medium", "high"]
 r"""Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response."""
 
 
@@ -763,7 +763,7 @@ r"""The version of photoReal to use. Must be v1 or v2. Only available for `leona
 CreatePromptEncodingFormat = Literal["float", "base64"]
 r"""The format to return the embeddings"""
 
-CreatePromptReasoningEffort = Literal["low", "medium", "high"]
+CreatePromptReasoningEffort = Literal["disable", "low", "medium", "high"]
 r"""Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response."""