+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -74,7 +74,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-06-24T11:53:53.481Z")
+    created: Optional[datetime] = parse_datetime("2025-06-24T19:47:49.832Z")
     r"""The date and time the resource was created"""