+++ packages/orq-rc/RELEASES.md
@@ -1839,3 +1839,13 @@ Based on:
 - [python v3.9.0-rc.5] packages/orq-rc
 ### Releases
 - [PyPI v3.9.0-rc.5] https://pypi.org/project/orq-ai-sdk/3.9.0-rc.5 - packages/orq-rc
+
+## 2025-06-24 19:49:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.568.2 (2.634.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.9.0-rc.6] packages/orq-rc
+### Releases
+- [PyPI v3.9.0-rc.6] https://pypi.org/project/orq-ai-sdk/3.9.0-rc.6 - packages/orq-rc